# server_django/openapi-schema.yml
openapi: 3.0.2
info:
  title: ''
  version: ''
paths:
...
  /search/{search}:
    get:
      operationId: retrievesearch
      description: ''
      parameters:
      - name: search
        in: path
        required: true
        description: ''
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: ''
      tags:
      - search
...